Q3 optim

+++ hw2/python/needle/ops/ops_mathematic.py
@@ -137,7 +137,7 @@ class DivScalar(TensorOp):
 
     def compute(self, a):
         ### BEGIN YOUR SOLUTION
-        return a / self.scalar
+        return (a / self.scalar).astype(a.dtype)
         ### END YOUR SOLUTION
 
     def gradient(self, out_grad, node):
